api/requirements.txt
uvicorn
fastapi
deta
python-dotenv
python-multipart
requests
aiofiles
boto3